Renamed laning page and component
Renamed the landing page and the form component contained within to better reflect their function.
