+++ client/src/styles/main-styles.js
@@ -8,7 +8,7 @@ const mainStyles = {
   },
   colorOverlay: {
     height: "100%",
-    background: "linear-gradient(180deg, #3A8DFF 0%, #86B9FF 100%)",
+    background: "linear-gradient(180deg, #3A8DFF 30%, #86B9FF 100%)",
     opacity: "85%",
     alignItems: "center",
     justifyContent: "center",
